requirements.txt
langchain
langchain-openai
langchain-community
langgraph
requests
beautifulsoup4
python-dotenv
tavily-python
pydantic
typing-extensions